all the requested commands are done in this commit exept the logout command wich will be commited soon this commit has some minor bug fixes for the pre existing functionalitys such as post mony and all that came with that command and the adding of a new command get money wich is fully tested

+++ movie.h
@@ -54,7 +54,7 @@ private:
 	int year;
 
 	int rater_count;
-	float rate;
+	int rate;
 
 	int sales_publisher;
